Refactor: delete tags.js and reallocate tag variables
