--- utils/tags.js
@@ -1,19 +1,0 @@
-const emailInput = document.querySelector(".email-input");
-const pwInput = document.querySelector(".password-input");
-const pwCheckInput = document.querySelector("#password-check-input");
-const pwWrapper = document.querySelector(".password-wrapper");
-const signinBtn = document.querySelector(".signin-btn");
-const eyeBtn = document.querySelector(".eye-off-btn");
-const pwCheckEyeBtn = document.querySelector(".pw-check-eye-off-btn");
-const signupBtn = document.querySelector(".signup-btn");
-
-export {
-  emailInput,
-  pwInput,
-  pwWrapper,
-  signinBtn,
-  eyeBtn,
-  pwCheckInput,
-  pwCheckEyeBtn,
-  signupBtn,
-};
